Fix review comments : javadoc
