+++ connect/api/src/main/java/org/apache/kafka/connect/rest/ConnectRestExtension.java
@@ -25,31 +25,34 @@ import java.io.Closeable;
 import java.util.Map;
 
 /**
- * A plugin interface to allow registration of new JAX-RS resources like Filters, REST endpoints,
- * providers, etc. The implementations will be discovered by the standard Java {@link
- * java.util.ServiceLoader} mechanism by  Connect's plugin class loading mechanism.
- * The implementation class must be packaged in a JAR that includes the
- * {@code META-INF/services/org.apache.kafka.connect.rest.extension.ConnectRestExtension} containing
- * the fully qualified name of the implementation class.
+ * A plugin interface to allow registration of new JAX-RS resources like Filters, REST endpoints, providers, etc. The implementations will
+ * be discovered using the standard Java {@link java.util.ServiceLoader} mechanism by  Connect's plugin class loading mechanism.
  *
- * <p> When Connect's worker configuration uses the REST extension implementation class,
- * upon startup Connect will instantiate the implementation and pass the configuration
- * to the instance via {@link Configurable#configure(Map)}.
+ * <p>The extension must be packaged as a plugin, with one JAR containing the implementation classes and a {@code
+ * META-INF/services/org.apache.kafka.connect.rest.extension.ConnectRestExtension} file that contains the fully qualified name of the
+ * class(es) that implement the ConnectRestExtension interface. The plugin should also include the JARs of all dependencies except those
+ * already provided by the Connect framework.
  *
- * <p> Typical use cases that can be implemented using this interface include things like security
- * (authentication and authorization), logging, request validations, etc.
+ * <p>To install into a Connect installation, add a directory named for the plugin and containing the plugin's JARs into a directory that is
+ * on Connect's {@code plugin.path}, and (re)start the Connect worker.
+ *
+ * <p>When the Connect worker process starts up, it will read its configuration and instantiate all of the REST extension implementation
+ * classes that are specified in the `rest.extension.classes` configuration property. Connect will then pass its configuration to each
+ * extension via the {@link Configurable#configure(Map)} method, and will then call {@link #register} with a provided context.
+ *
+ * <p>When the Connect worker shuts down, it will call the extension's {@link #close} method to allow the implementation to release all of
+ * its resources.
  */
 public interface ConnectRestExtension extends Configurable, Versioned, Closeable {
 
     /**
-     * ConnectRestExtension implementation can register custom JAX-RS resources via the {@link
-     * #register(ConnectRestExtensionContext)} method. The Connect Framework will invoke this method
-     * after registering the default Connect resources. If the implementations attempt to
-     * re-register any of the Connect resources, it will be be ignored and will be logged.
+     * ConnectRestExtension implementation can register custom JAX-RS resources via the {@link #register(ConnectRestExtensionContext)}
+     * method. The Connect Framework will invoke this method after registering the default Connect resources. If the implementations attempt
+     * to re-register any of the Connect resources, it will be be ignored and will be logged.
      *
-     * @param restPluginContext The context provides access to JAX-RS {@link javax.ws.rs.core.Configurable}
-     *                          and {@link ConnectClusterState}.The custom JAX-RS resources can be
-     *                          registered via the {@link ConnectRestExtensionContext#configurable()}
+     * @param restPluginContext The context provides access to JAX-RS {@link javax.ws.rs.core.Configurable} and {@link
+     *                          ConnectClusterState}.The custom JAX-RS resources can be registered via the {@link
+     *                          ConnectRestExtensionContext#configurable()}
      */
     void register(ConnectRestExtensionContext restPluginContext);
 }
